Operational State rename
