+++ src/app/clusters/operational-state-server/app_config_dependent_sources.cmake
@@ -16,7 +16,9 @@
 TARGET_SOURCES(
   ${APP_TARGET}
   PRIVATE
+    "${CLUSTER_DIR}/OperationalStateCluster.cpp"
+    "${CLUSTER_DIR}/OperationalStateCluster.h"
+    "${CLUSTER_DIR}/OperationalStateDelegate.h"
     "${CLUSTER_DIR}/operational-state-cluster-objects.h"
-    "${CLUSTER_DIR}/operational-state-server.cpp"
     "${CLUSTER_DIR}/operational-state-server.h"
 )
